+++ packages/pangea-sdk/tests/integration/__init__.py
@@ -1,7 +1,7 @@
-from .test_audit import *
-from .test_authn import *
-from .test_embargo import *
-from .test_file_scan import *
-from .test_intel import *
-from .test_redact import *
-from .test_vault import *
+from .test_audit import TestAudit
+from .test_authn import TestAuthN
+from .test_embargo import TestEmbargo
+from .test_file_scan import TestFileScan
+from .test_intel import TestDomainIntel, TestFileIntel, TestIPIntel, TestURLIntel, TestUserIntel
+from .test_redact import TestRedact
+from .test_vault import TestVault