+++ CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Audit /download_results endpoint support
+
 ### Fixed
 
 - AuthN list method's filter not being serialized properly.
 - Incorrect docstrings positioning in `PangeaConfig`.
 - Incorrect variable name in `Vault` docstring.
 - Extraneous colons in `Vault.key_rotate()` docstrings.
+
+### Changed
+
+- AuthN ClientTokenCheckResult `token` field is optional
 
 
 ## [3.7.0] - 2024-02-26